+++ packages/syllst/src/index.ts
@@ -29,4 +29,4 @@
  */
 
 // Export all types
-export * from './types';
+export * from './types/index.js';