+++ packages/validator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @syllst/validator
 
+## 0.4.1
+
+### Patch Changes
+
+- @syllst/processor@0.4.1
+
 ## 0.4.0
 
 ### Minor Changes